backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.2",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "ruff>=0.12.2",
]
[project.scripts]
format = "ruff format ."
lint = "ruff check ."
typecheck = "mypy src/"

[tool.ruff]
line-length = 88
target-version = "py313"
[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]  # line too long

[tool.pyright]
include = ["."]
exclude = ["**/__pycache__"]
ignore = ["**/node_modules"]
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.13"

